Maze + pacman movement
Still needs to be transferred to the main project. D-pad is temporary and needs to be replaced by finished d-pad. Ghost does not move and needs to be implemented. Pacman snaps from space to space rather than moving smoothly. Pellets do not add to a score. Power pellets do not activate any powerups.